fix :修复压测启动报错的问题

+++ api/src/main/java/io/shulie/flpt/pressure/engine/api/entity/EnginePressureConfig.java
@@ -42,7 +42,7 @@ public class EnginePressureConfig extends AbstractEntry {
     /**
      * 循环次数
      */
-    private Integer loopsNum;
+    private Long loopsNum;
     /**
      * 采样率
      */
